refactor(storage): localize projection tests
